Update okio
By declaring it as a direct dependency.

This is an attempt to avoid using a vulnerable version of
com.squareup.okio:okio by adding it as a direct dependency, taking
it from 1.17.2 (as resolved in the dependency graph, though I'm not
sure how often that was really the version being used in practice)
to 3.5.0.

Versions prior to 3.4.0 are affected by CVE-2023-3635.